requirements.txt
openai==0.28.1
flask==3.0.0
python-dotenv==1.0.0
langchain==0.0.308
matplotlib==3.8.2
plotly==5.18.0
pandas==2.1.1
scipy==1.11.3
scikit-learn==1.3.1
tiktoken==0.5.2
faiss-cpu==1.7.4
constants==0.6.0
transformers

# 这个有点大(670.2 MB),看自己的需求
torch